requirements.txt
azure-functions
msrestazure==0.6.4
azure-identity==1.5.0
azure-mgmt-netapp>=0.6.0
azure-mgmt-resource>=3.1.0
